Well, after much scabbling around for clues, hacking and some help, it's here.
This is a basic functional Apache module. It needs significant cleaning up
which I'll start doing in the next few updates. I thought it'd be useful for
others who follow to see some of the warts in the code as they indicate areas
where I, and you, may struggle.

There are known limitations:
- Little consideration for concurrency
- It still has some references to example_module rather than kiwix_module
- It's limited to a single ZIM file
- It won't work with older ZIMs that have hard-coded, full internal paths.
Newer ZIM files use relative paths.

But, it works! :)