Added the ability to configure the path and filename of the ZIM file.
Another milestone achieved. Beware, at least based on my experience the
Apache Server is very sensitive to the contents of the configuration file and
can refuse to start `sudo service apache2 restart` if it doesn't like the
contents of the file. I'll learn and experiment to try and determine the safe
boundaries for the file's contents. Meanwhile I'll update the README next.

Added a `./status` feature
Added a `./config` feature

These and others can be enhanced e.g. with usage statistics.

Thank you to Mattieu in particular for the cunning workaround to access the
configuration data despite problems with the Apache macro being problematic in
C++